use gin framework to implement the API, change database table column type, test api with dummy data after fixing some git issue

+++ app/schema.sql
@@ -3,23 +3,13 @@ DROP TABLE IF EXISTS accounts;
 
 CREATE TABLE IF NOT EXISTS transactions (
     id TEXT NOT NULL PRIMARY KEY,
-    account_id INTEGER NOT NULL, 
-    amount INTEGER,
+    account_id REAL NOT NULL, 
+    amount REAL,
     created_at TEXT,
     FOREIGN KEY(account_id) REFERENCES accounts(id)
 );
 
 CREATE TABLE IF NOT EXISTS accounts (
     Id TEXT NOT NULL PRIMARY KEY,
-    balance INTEGER
+    balance REAL
 );
-
--- INSERT INTO Categories (Id, Name) VALUES
--- (1, "Watersports"),database
--- (2, "Soccer");
-
--- INSERT INTO Products (Id, Name, Category, Price) VALUES 
--- (1, "Kayak", 1, 279),
--- (2, "Lifejacket", 1, 48.95),
--- (3, "Soccer Ball", 2, 19.50),
--- (4, "Corner Flags", 2, 34.95);
